Remove duplicate loading system class
