+++ packages/client-core/src/world/ClientModule.ts
@@ -29,7 +29,7 @@ import { OverlaySystem } from '@ir-engine/client-core/src/systems/OverlaySystem'
 import { AvatarSpawnSystem } from '../networking/AvatarSpawnSystem'
 import { AvatarUISystem } from '../systems/AvatarUISystem'
 import { LinkErrorSystem } from '../systems/LinkErrorSystem'
-import { LoadingUISystem } from '../systems/LoadingUISystemBasic'
+import { LoadingUISystem } from '../systems/LoadingUISystem'
 import { MediaControlSystem } from '../systems/MediaControlSystem'
 import { PositionalAudioSystem } from '../systems/PositionalAudioSystem'
 import { PoiUiSystem } from '../systems/ui/PoiUI'
